[BISERVER-8242] - Support for metadata injection of Shared Dimension step.
