+++ src/main/resources/org/pentaho/di/trans/steps/annotation/messages/messages_en_US.properties
@@ -193,3 +193,9 @@ AnnotateStream.Injection.CALC_MEASURE_DESCRIPTION=This is a brief description of
 AnnotateStream.Injection.CALC_MEASURE_IS_HIDDEN=This option will hide this calculated measure in the data model.
 AnnotateStream.Injection.CALC_MEASURE_CALCULATE_SUBTOTALS=This option will allow your calculated measure to be used in calculations of subtotals in your reports.
 AnnotateStream.Injection.CALC_MEASURE_FORMULA=The formula of your calculated measure.
+
+AnnotateStream.Injection.SHARED_DIMENSION_NAME=The name of the shared dimension.
+AnnotateStream.Injection.DATA_PROVIDER_STEP=Choose a source to associate your dimension from. The source must be a valid step.
+AnnotateStream.Injection.DIMENSION_KEY=Create Dimension Key
+AnnotateStream.Injection.DIMENSION_KEY_DIMENSION=This option lets you specify the dimension or create a new dimension for the key. This is a required field.
+AnnotateStream.Injection.DIMENSION_KEY_FIELD=The name of the field to be annotated.
